chore(core): split out code in state_processor.go and state_processor_test.go - Move our code from state_processor.go to state_processor_ext.go - Move our code from state_processor_test.go to state_procesor_ext_test.go - Slight modifications to state_processor.go to make it more alike geth master
